修复：一些bbox是int类型

+++ adet/data/detection_utils.py
@@ -77,8 +77,9 @@ def annotations_to_instances(annos, image_size, mask_format="polygon"):
     # add global attention
     attention = np.stack(
             [BoxMode.convert(obj["bbox"][None, ...], obj["bbox_mode"], BoxMode.XYWH_ABS).squeeze() for obj in annos])
-    # 新增标注.此处假设segm都只含有一个多边形
-    attention_mask = torch.from_numpy(np.ascontiguousarray(polygons_to_bitmask(attention, *image_size)))
+    # 新增标注.此处假设segm都只含有一个多边形; 在一些标注中bbox是int类型，所以需要转化
+    attention_mask = torch.from_numpy(np.ascontiguousarray(polygons_to_bitmask(attention.astype(float),
+                                                                               *image_size)))
     attentions = BitMasks(
         torch.stack([attention_mask] + [torch.zeros_like(attention_mask) for _ in range(len(annos)-1)])
     )
